libvirt: new handler
Libvirt handler connects to running libvirt daemon and fetches
information about libvirt domains in XML. It is then parsed to fill
internal vm structures.

Adds dependency on expat and libvirt. Expat itself is a dependency of
e.g. avahi, so it should be everywhere. There are plans to enable user
to compile plotnetcfg with these features disabled, releasing both
dependencies.

Signed-off-by: Ondrej Hlavaty <ohlavaty@redhat.com>